Fixed undefined behavior when using TLS
The mutable reference created by UnsafeCell stayed alive across
the recursive call to inner_fn_call, which resulted in a mutable alias.

To ensure that this kind of issue results in a panic rather than
undefined behavior, I've switched to using a RefCell instead of an
UnsafeCell. The RefMut obtained from the RefCell is explicitly dropped
before we make the recursively call, ensuring that at most one is alive
at any time.